pyproject.toml
[project]
name = "selfrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
 "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.0",
    "langchain-pinecone==0.2.13",
    "langchain-text-splitters>=1.1.1",
    "pinecone>=6.0.0,<8.0.0",
    "pypdf>=6.7.1",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.3",
    "redis>=5.0.0",
    "numpy>=1.26.0",
]
